Release 1.2.0 (#21)
* fix: :bug: Get good order for votes (#16)

The votes are now retrieved by sorting them with -begin_date

* fix: :bug: Fix number of voters on results pages (#17)

Fix number of voters on results page. Close #17.

* docs: :memo: Improve docstr coverage (#12)

Improve docstr coverage to 100%. Close issue #12.

* fix: :bug: Fix translation of groups on the votes index (#15)

Close issue #15.

* feat: :sparkles: Add field to know who voted (#18)

* feat: :sparkles: Add field to know who voted (#18)

This creates a boolean field to set if voters should be seen on the results page. This field cannot be modified after creation

* fix: :bug: Remove mention of VOTE_SEE_VOTERS

* fix: :fire: Remove useless import

* feat: :art: Add logo (#14)

Add logo and add it in favicon and on the login page. Close #14

* feat: :sparkles: Allow votes to have public results

Fix #19

* chore: Bump 1.2.0